+++ src/main/java/ru/kata/spring/boot_security/demo/services/UserService.java
@@ -1,9 +1,8 @@
 package ru.kata.spring.boot_security.demo.services;
 
 
-
-
 import ru.kata.spring.boot_security.demo.model.User;
+
 import java.util.List;
 
 public interface UserService {
@@ -12,9 +11,9 @@ public interface UserService {
 
     public void saveUser(User user);
 
-    public User getUser(int id);
+    public User getUser(Long id);
 
-    public void deleteUser(int id);
+    public void deleteUser(Long id);
 
     public User findByUsername(String username);
 }
